Create QueryClient
This program takes a student's number as an argument. This program will send the student's number as a request to the server and wait for the server's reply. This program will display the student's contact details once it receives from the server.
This program uses UDP as the communication protocol to program the client and the server interaction.