Create QueryServer
This program is continuously waiting for requests from clients. Once the server receives a request, it will create a thread to process the request and send a response to the client. It means that the server follows the thread per request multi threading model. The server maintains student's contact details in a table in a database. Each time the server receives a request, it will query the table for the corresponding student's contact details.